feat: created scheduler to run data backup job when starting web server
