+++ run.sh
@@ -8,5 +8,5 @@
 #
 
 . ./env.sh
-python manage.py backupfiles &
+python manage.py run_data_backup &
 gunicorn -b :$PORT -w 2 --timeout 90 localcontexts.wsgi:application
